requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
alembic==1.13.2
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.4.0
certifi==2025.11.12
et_xmlfile==2.0.0
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.4
openpyxl==3.1.5
pandas==2.2.3
pillow==12.0.0
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
qrcode==7.4.2
six==1.17.0
SQLAlchemy==2.0.35
typing_extensions==4.15.0
tzdata==2025.2
yarl==1.22.0
